todoAgent/pyproject.toml
[project]
name = "todoagent"
version = "0.1.0"
description = "AI-powered Todo Agent using OpenAI Agents SDK with Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "chainlit>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
